Update animator
This change reworks the animator to utilize quaternions for rotations and adds scaling support. The part that modifies the display entities has true/false condition checks to allow external sources to modify the parts while the model is animating. I removed delay_pose as I don't see the point of using that since the sync is pretty much perfect.