Update the loader for DModels
This change reduces the scale factor to allow bigger models to load correctly due to Minecraft item size limitations still applying. The pack version has been increased to 13. For the rotation channel the euler angles are converted to a quaternion I made a procedure below that does this however I do recommend implementing something like it for the quaternion tag this is also done for the outline rotation as well.
I added tintindex to allow colorable models.